Migrated satellite helper methods to util file
